python/requirements.txt
web3==6.20.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
eth-abi==5.1.0
eth-account==0.11.2
requests==2.32.3
pytest==8.2.2
hypothesis==6.111.0
